refactor(wasm): split trace-snapshot accessors into wasm/traces.rs
Start splitting the ~2470-line impl WasmSimulator in wasm/lib.rs. First:
the trace-snapshot accessors (air/uart/wifi/fdcan/iolink + drain_uart_output)
move to wasm/traces.rs as a second #[wasm_bindgen] impl block — child module
sees the crate-root struct's private fields, and wasm_bindgen supports
multiple impl blocks.

lib.rs 2540 -> 2392. Verified: wasm32-unknown-unknown + host builds, clippy clean.
